# library_distributed_server/docs/coordinator/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Coordinator service for distributed transactions in the library management system",
        "title": "Distributed Transaction Coordinator API",
        "termsOfService": "N/A",
        "contact": {
            "name": "API Support",
            "email": "dev@library.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user using stored procedure and return JWT access token only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful with access token",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logout user (invalidate token on client side)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about the currently authenticated user including role, branch, and permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "User profile information",
                        "schema": {
                            "$ref": "#/definitions/models.UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/coordinator/transfer-book": {
            "post": {
                "description": "Transfer a book copy from one site to another using distributed transaction coordination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coordinator"
                ],
                "summary": "Transfer book between sites using 2PC",
                "parameters": [
                    {
                        "description": "Book transfer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TransferBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book transferred successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TransferBookResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to transfer book",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginRequest": {
            "description": "User login credentials",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "thuthu01"
                }
            }
        },
        "auth.LoginResponse": {
            "description": "Successful login response with JWT access token only",
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "JWT access token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "main.TransferBookRequest": {
            "type": "object",
            "required": [
                "fromSite",
                "maQuyenSach",
                "toSite"
            ],
            "properties": {
                "fromSite": {
                    "description": "Source site ID",
                    "type": "string",
                    "example": "Q1"
                },
                "maQuyenSach": {
                    "description": "Book copy ID to transfer",
                    "type": "string",
                    "example": "QS001"
                },
                "toSite": {
                    "description": "Destination site ID",
                    "type": "string",
                    "example": "Q3"
                }
            }
        },
...
        "models.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "details": {
                    "description": "Error details (optional)"
                },
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
...
        "models.TransferBookResponse": {
            "description": "Response after successful book transfer using 2PC protocol",
            "type": "object",
            "properties": {
                "coordinator": {
                    "description": "Coordinator service",
                    "type": "string",
                    "example": "Distributed Transaction Coordinator"
                },
                "fromSite": {
                    "description": "Source site ID",
                    "type": "string",
                    "example": "Q1"
                },
                "maQuyenSach": {
                    "description": "Transferred book copy ID",
                    "type": "string",
                    "example": "QS001"
                },
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Book transferred successfully using 2PC protocol"
                },
                "protocol": {
                    "description": "Protocol used",
                    "type": "string",
                    "example": "Two-Phase Commit (2PC)"
                },
                "toSite": {
                    "description": "Destination site ID",
                    "type": "string",
                    "example": "Q3"
                }
            }
        },
        "models.UserInfo": {
            "description": "Detailed user information including permissions",
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "id": {
                    "description": "User ID",
                    "type": "string",
                    "example": "user123"
                },
                "maCN": {
                    "description": "Branch code (for THUTHU)",
                    "type": "string",
                    "example": "Q1"
                },
                "permissions": {
                    "description": "User permissions",
                    "type": "string",
                    "example": "SELECT,INSERT,UPDATE,EXECUTE"
                },
                "role": {
                    "description": "User role",
                    "type": "string",
                    "enum": [
                        "THUTHU",
                        "QUANLY"
                    ],
                    "example": "THUTHU"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "ThuThu_Q1"
                }
            }
        }
    }
}